all: Move most action functions into their dialogs
The exception is ForceRemoveModal, which is used for different kinds
of actions.
